Add files via upload
Final Creative Project showcasing animated Christmas tree where I created code to create data tree plot, made the points, and used the "Jigglr" package to make the snow to fall on the tree plot. I used the ggplot package to make the data tree plot and created a data frame for each color of the ornaments in order to create points on the Christmas tree. For the snow points, I also used ggplot and coded to create a star-shape and to make the snow "fall", I made a code loop within the "Jigglr" for it to reach certain points in a vertical manner.